# fix(perf): enable compile cache